backend/requirements.txt
flask
flask-cors
flask-jwt-extended
pymysql
python-dotenv
passlib
pyjwt
lightgbm
pandas
numpy
scikit-learn
google-genai
gunicorn
bcrypt
